[ONNX] Separate fx _type_utils from torchscript exporter
[ghstack-poisoned]
